requirements.txt
cffi==1.11.2
cycler==0.10.0
lxml==4.2.5
matplotlib==2.1.1
numpy==1.15.4
PyAudio==0.2.11
pycparser==2.18
pyparsing==2.2.0
PySoundFile==0.9.0.post1
python-dateutil==2.6.1
python-toolbox==0.9.3
pytz==2017.3
scipy==1.1.0
six==1.11.0
